+++ internal/api/bulking/handler_stream_text_test.go
@@ -75,7 +75,7 @@ send [USD 100] (
 				select {
 				case receive <- BulkElementResult{
 					Data:      ledger.CreatedTransaction{},
-					LogID:     id + 1,
+					LogID:     uint64(id) + 1,
 					ElementID: id,
 				}:
 				case <-time.After(100 * time.Millisecond):
